+++ src/app/header/header-searchfield-flex-demo.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+    selector: 'app-searchfield-flex-header-demo',
+    templateUrl: 'header-searchfield-flex.demo.html'
+})
+export class HeaderSearchfieldFlexDemoComponent {
+}