+++ src/app/header/header-searchfield.demo.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+    selector: 'app-searchfield-header-demo',
+    templateUrl: 'header-searchfield.demo.html'
+})
+export class HeaderSearchfieldDemoComponent {
+}